detector/requirements.txt
requests==2.31.0
scikit-learn==1.3.2
numpy==1.26.1
python-dateutil==2.8.2

